% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRFDrawer.R
\name{createRFDrawer}
\alias{createRFDrawer}
\title{UI and Server code returner}
\usage{
createRFDrawer(defaultColor = "blue", defaultShape = "square",
  colors = NULL, shapes = NULL)
}
\arguments{
\item{defaultColor}{The default color of the class.}

\item{defaultShape}{The default shape of the class.}

\item{colors}{The colors for the classes.}

\item{shapes}{The shapes for the classes.}
}
\value{
With a list which contains the UI and the Server code.
}
\description{
Returns with a list which contains the called ui and server function.
}
