% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRFDrawerServer.R
\name{createRFDrawerServer}
\alias{createRFDrawerServer}
\title{Returns with the random forest drawer server closure}
\usage{
createRFDrawerServer(defaultColor, defaultShape, colors, shapes)
}
\arguments{
\item{defaultColor}{The default color of the class.}

\item{defaultShape}{The default shape of the class.}

\item{colors}{The colors for the classes.}

\item{shapes}{The shapes for the classes.}
}
\value{
With the server closure.
}
\description{
This function returns with a server closure which can be passed to eg. callModule.
}
