% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRFDrawerUI.R
\name{createRFDrawerUI}
\alias{createRFDrawerUI}
\title{Returns with the random forest drawer UI closure}
\usage{
createRFDrawerUI()
}
\value{
With the UI closure.
}
\description{
This function returns with an UI closure which can be passed to eg. renderUI.
}
