% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColorForClass.R
\name{getColorForClass}
\alias{getColorForClass}
\title{Get color for a class}
\usage{
getColorForClass(classValue, defaultColor = "blue", classes = NULL,
  colors = NULL)
}
\arguments{
\item{classValue}{The caller wants a color for this class.}

\item{defaultColor}{The default color, if we can't return with a color.}

\item{classes}{A list of possibly classes.}

\item{colors}{A list which contains the colors of the classes.}
}
\value{
A color as a string.
}
\description{
Get color for a factor's class based on the parameter values.
}
\details{
This is basically a LUT, which finds the proper color for a factor's class.
We can add the \code{classes}, the \code{colors} or a \code{defaultColor}.
}
\examples{
\dontrun{
print(getColorForClass("Test",classes = c("Test", "Test2"), colors=c("red", "blue")))
print(getColorForClass("Test2"))
print(getColorForClass(NA,classes = c("Test", "Test2"), colors=c("red", "blue")))
}
}
