% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShapeForClass.R
\name{getShapeForClass}
\alias{getShapeForClass}
\title{Get shape for a class}
\usage{
getShapeForClass(classValue, defaultShape = "square", classes = NULL,
  shapes = NULL)
}
\arguments{
\item{classValue}{The caller wants a shape for this class.}

\item{defaultShape}{The default shape, if we can't return with a shape}

\item{classes}{A list of possibly classes.}

\item{shapes}{A list which contains the shapes of the classes.}
}
\value{
A shape as a string.
}
\description{
Get shape for a factor's class based on the parameter values.
}
\details{
This is basically a LUT, which finds the proper shape for a factor's class.
We can add the \code{classes}, the \code{shapes} or a \code{defaultShape}.
}
\examples{
\dontrun{
print(getShapeForClass("Test",classes = c("Test", "Test2"), shapes=c("triangle", "circle")))
print(getShapeForClass("Test2"))
print(getShapeForClass(NA,classes = c("Test", "Test2"), shapes=c("triangle", "circle")))
}
}
