% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVisNetworkFromIGraph.R
\name{getVisNetworkFromIGraph}
\alias{getVisNetworkFromIGraph}
\title{Get visNetwork from IGraph wrapper}
\usage{
getVisNetworkFromIGraph(graph)
}
\arguments{
\item{graph}{The IGraph graph which we want to convert.}
}
\value{
A fully functional visNetwork
}
\description{
Converts and IGraph graph into a visNetwork.
}
\details{
Converts an IGraph into a visNetwork which has a hiearchical layout.
Made some performance boosting.
But it's still interactive.
This function also makes the nice tree layout for the graph.
}
